Smart Resume Screening System Using NLP and Machine Learning/requirements.txt
streamlit
pdfplumber
python-docx
nltk
scikit-learn
pypdfium2
